style: add `phosphor-icons/react` icon package
Integrates @phosphor-icons/react package, the latest version of the Phosphor Icons.

Previously the project was using icons from the legacy version, phosphor-react, which was a dependency of the keep-react UI library in the project.
However, the keep-react was removed in the previous commit(f37cdef). So decided to add latest version package

[Phosphor Icons Library Website: https://phosphoricons.com/]
[NPM package page link: https://www.npmjs.com/package/@phosphor-icons/react]